working on navigate
